seperated graph into its own component and add/remove graph
